refactor: QueryBuilder 삭제
