+++ src/main/java/database/sql/dml/CustomSelect.java
@@ -3,6 +3,7 @@ package database.sql.dml;
 import database.mapping.Association;
 import database.mapping.EntityMetadata;
 import database.mapping.EntityMetadataFactory;
+import database.sql.dml.part.WhereClause;
 
 import java.util.List;
 import java.util.Map;
