+++ src/main/java/database/sql/dml/part/where/OperatorType.java
@@ -1,4 +1,4 @@
-package database.sql.dml.where;
+package database.sql.dml.part.where;
 
 enum OperatorType {
     IS("IS"),
